restructure to package search in a separate directory
