+++ musclex/utils/bg_search/__init__.py
@@ -1,0 +1,5 @@
+"""Background search helpers for iterative 2D background fitting.
+
+Holds the full-image search routines (``background_search``) and
+their low-level utilities (``background_search_utils``).
+"""
